# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.CreateCorrectionRequest:
    properties:
      corrected_value:
        type: string
      field_name:
        type: string
    required:
    - corrected_value
    - field_name
    type: object
...
  models.CorrectionResponse:
    properties:
      corrected_value:
        type: string
      created_at:
        type: string
      field_name:
        type: string
      id:
        type: string
      original_value:
        type: string
      reviewed_at:
        type: string
      scan_id:
        type: string
      status:
        $ref: '#/definitions/models.CorrectionStatus'
    type: object
  models.CorrectionStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - CorrectionStatusPending
    - CorrectionStatusApproved
    - CorrectionStatusRejected
...
  response.ErrorDetail:
    description: Error details
    properties:
      code:
        example: 400
        type: integer
      field:
        example: email
        type: string
      message:
        example: Invalid request
        type: string
    type: object
  response.ErrorEnvelope:
    description: Standard error response envelope
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
      errors:
        items:
          $ref: '#/definitions/response.ErrorDetail'
        type: array
      metadata:
        $ref: '#/definitions/response.Metadata'
      success:
        example: false
        type: boolean
    type: object
  response.Metadata:
    description: Response metadata
    properties:
      request_id:
        example: abc123
        type: string
      timestamp:
        example: "2024-12-07T10:00:00Z"
        type: string
      version:
        example: 1.0.0
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: habbazahubbal@gmail.com
    name: Hubbal Kholiq Habbaza
  description: API untuk memproses foto nutrition facts dan barcode
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: NutriSnap API
  version: 1.0.0
paths:
...
  /scan/{id}/correct:
    post:
      consumes:
      - application/json
      description: User can submit corrections for incorrect OCR data
      parameters:
      - description: Scan ID
        in: path
        name: id
        required: true
        type: string
      - description: Correction data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateCorrectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CorrectionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Submit a correction for scan data
      tags:
      - Correction
  /scan/{id}/corrections:
    get:
      consumes:
      - application/json
      description: Get all corrections submitted for a specific scan
      parameters:
      - description: Scan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CorrectionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get corrections for a scan
      tags:
      - Correction
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"